requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audeer==2.2.2
audformat==1.3.2
audinterface==1.3.1
audiofile==1.5.1
audmath==1.4.2
audobject==0.7.12
audresample==1.3.4
cffi==1.17.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.0
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
importlib_metadata==8.7.0
iso3166==2.1.1
iso639-lang==2.6.3
joblib==1.5.1
kiwisolver==1.4.9
matplotlib==3.10.5
multidict==6.6.3
numpy==2.2.6
opensmile==2.5.1
oyaml==1.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
playwright==1.53.0
praat-parselmouth==0.4.6
propcache==0.3.2
pyarrow==21.0.0
pycparser==2.22
pydub==0.25.1
pyee==13.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
soundfile==0.13.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
yarl==1.20.1
zipp==3.23.0
